Introduces section components to make the code more readable
